Fixed column type for external key.

+++ src/ututi/migration/189_i18n_tables_upgrade.sql
@@ -4,7 +4,7 @@ create table i18n_texts (
 
 
 create table i18n_texts_versions (
-       i18n_texts_id bigserial not null references i18n_texts(id) on delete cascade,
+       i18n_texts_id int8 not null references i18n_texts(id) on delete cascade,
        language_id varchar(100) not null references languages(id) on delete cascade,
        text text not null default '',
        primary key (i18n_texts_id, language_id));;
